Fix coverage targets by allowing ow2-asm
There was a recent change (ag/35808958) for jacoco to not use the
prebuilt asm dependency anymore, but switch to the dependency that can be built from source, ow2-asm. However, this dependency was not yet part of the allowed_deps.txt.

Bug: 446603650
Test: Build a coverage target to make sure that it succeeds
Flag: TEST_ONLY
Apex-Size-Increase: None, as it is replacing a different
 dependency & it will only be used as part of Jacoco in
 coverage builds

Change-Id: Ic53b3fc0e8442e032efcd7831fe8033872d35f76